refactor: Updating frontend file structure
