+++ frontend/src/App.jsx
@@ -1,29 +1,30 @@
-
 import React, {useState} from "react";
 import './App.css';
-import {Login} from "./components/pages/Login";
 import { Routes, Route, BrowserRouter } from 'react-router-dom'
 
 /* Web Pages */
-import Calendar from './components/pages/Calendar'
-import Flashcards from './components/pages/Flashcards'
-import Home from './components/pages/Home'
-import Notes from './components/pages/Notes'
-import Todo from './components/pages/Todo'
-import RegisterForm from './components/pages/RegisterForm'
+import Login from "./pages/Login";
+import Signup from "./pages/Signup";
+
+import Calendar from "./pages/app/Calendar";
+import Flashcards from "./pages/app/Flashcards";
+import Home from "./pages/app/Home";
+import Notes from "./pages/app/Notes";
+import Todo from "./pages/app/Todo";
 
 function App() {
   return (    
     <BrowserRouter>
       <Routes>
-          <Route path= "/login" element={<Login/>}></Route>
           <Route path="/" element={<Home/>}></Route>
-          <Route path="/app/home" element={<Home/>}></Route>
-          <Route path="/app/todo" element={<Todo/>}></Route>
+          <Route path="/login" element={<Login/>}></Route>
+          <Route path="/signup" element={<Signup/>}></Route>
+
           <Route path="/app/calendar" element={<Calendar/>}></Route>
           <Route path="/app/flashcards" element={<Flashcards/>}></Route>
+          <Route path="/app/home" element={<Home/>}></Route>
           <Route path="/app/notes" element={<Notes/>}></Route>
-          <Route path="/app/RegisterForm" element={<RegisterForm/>}></Route>
+          <Route path="/app/todo" element={<Todo/>}></Route>
         </Routes>
     </BrowserRouter>
   )
