+++ frontend/src/__tests__/Calendar.test.jsx
@@ -1,4 +1,4 @@
-import Calendar from '../components/pages/Calendar';
+import Calendar from '../pages/app/Calendar';
 import {render,screen} from '@testing-library/react';
 
 test("should have hello world", () => {
