--- frontend/src/components/BasicTemplateComponent.jsx
@@ -1,7 +1,0 @@
-const BasicTemplateComponent = () => {
-  return (
-    <div>BasicTemplateComponent</div>
-  )
-}
-
-export default BasicTemplateComponent
