--- frontend/src/components/index.jsx
@@ -1,1 +1,0 @@
-export { default as BasicTemplateComponent } from './BasicTemplateComponent';
