--- frontend/src/components/sum.jsx
@@ -1,4 +1,0 @@
-// sum.js
-export function sum(a, b) {
-    return a + b
-}
