--- frontend/src/pages/BasicTemplatePage.jsx
@@ -1,7 +1,0 @@
-const BasicTemplatePage = () => {
-  return (
-    <div>BasicTemplatePage</div>
-  )
-}
-
-export default BasicTemplatePage
