+++ frontend/src/pages/app/Flashcards.jsx
@@ -1,4 +1,4 @@
-import Sidebar from "../sidebar/Sidebar"
+import Sidebar from "../../components/sidebar/Sidebar"
 
 export default function Flashcards() {
     return (
