+++ frontend/src/pages/app/Home.jsx
@@ -1,4 +1,4 @@
-import Sidebar from "../sidebar/Sidebar"
+import Sidebar from "../../components/sidebar/Sidebar"
 
 
 export default function Home() {
