+++ frontend/src/pages/app/Todo.jsx
@@ -1,10 +1,13 @@
-import Sidebar from "../sidebar/Sidebar"
-import './TodoComponents/TodoApp.css'
-import TodoList from './TodoComponents/TodoList'
 /*Created with the Help of Brian Degins video and recourses
     YouTube Channel : https://www.youtube.com/@briandesign
     Video: https://www.youtube.com/watch?v=E1E08i2UJGI&list=LL&index=3
-    */
+*/
+
+import Sidebar from "../../components/sidebar/Sidebar"
+import TodoList from '../../components/todo/TodoList'
+
+import '../../style/TodoApp.css'
+
 export default function Todo() {
     return (
         <div className="flex">
